Remove AuthenticatedUser class and file, use User type instead
