+++ sso-kit-client/sso-kit-client-lit/README.md
@@ -1,8 +1,8 @@
-# @sso-kit/sso-kit-client
+# @hilla/sso-kit-client-lit
 
 The SSO Kit Client is the client side dependency for the SSO Kit for Hilla. It depends on the [sso-kit-starter-hilla](https://github.com/vaadin/sso-kit/tree/main/sso-kit-starter-hilla).
 
-Getting Started with Lit [documentation](https://hilla.dev/docs/lit/acceleration-kits/sso-kit/getting-started/#frontend).
+Getting Started with the SSO Kit for Hilla using Lit [documentation](https://hilla.dev/docs/lit/acceleration-kits/sso-kit/getting-started/#frontend).
 
 ## Installation
 
@@ -12,12 +12,10 @@ Install the client:
 npm i @hilla/sso-kit-client-lit@2.1.0
 ```
 
-Once installed, you can import the `SingleSignOnContext` and the `AuthenticatedUser` classes in your application:
+Once installed, you can import the `SingleSignOnContext` class in your application:
 
 ```js
-import { singleSignOnContext } from '@hilla/sso-kit-client-lit/SingleSignOnContext.js';
-
-import { authenticatedUser } from '@hilla/sso-kit-client-lit/AuthenticatedUser.js';
+import { singleSignOnContext as sso} from '@hilla/sso-kit-client-lit/SingleSignOnContext.js';
 ```
 
 ## Development
