+++ sso-kit-client/sso-kit-client-lit/package.json
@@ -29,9 +29,6 @@
     },
     "./SingleSignOnContext.js": {
       "default": "./SingleSignOnContext.js"
-    },
-    "./AuthenticatedUser.js": {
-      "default": "./AuthenticatedUser.js"
     }
   },
   "author": "Vaadin Ltd",
